feat: SMALLINT, INT support (#600)
* smallint, tests

* dory cpu helper

* SMALLINT and INT

* update dory helper with types

* types and tests

* diff

* tests and query context builder

* tests, docs, rebase, refactor

* is_integer method

* fmt

* Update crates/proofs/src/sql/ast/group_by_util.rs


* Update crates/proofs/src/sql/ast/group_by_util.rs


* Update crates/proofs/src/base/database/test_accessor_utility.rs


* Update crates/proofs/src/base/commitment/committable_column.rs


* polars

* rebase, owned table update

* doctest

* matching and naming, main rebase

* refactor

* update to main

---------

+++ crates/proofs/tests/integration_tests.rs
@@ -319,8 +319,8 @@ fn we_can_prove_a_complex_query_with_dory() {
     accessor.add_table(
         "sxt.table".parse().unwrap(),
         owned_table([
-            bigint("a", [1, 2, 3]),
-            bigint("b", [1, 0, 1]),
+            smallint("a", [1_i16, 2, 3]),
+            int("b", [1, 0, 1]),
             bigint("c", [3, 3, -3]),
             bigint("d", [1, 2, 3]),
             varchar("e", ["d", "e", "f"]),
@@ -348,8 +348,8 @@ fn we_can_prove_a_complex_query_with_dory() {
         .unwrap()
         .table;
     let expected_result = owned_table([
-        bigint("a", [1]),
-        bigint("b", [1]),
+        smallint("a", [1_i16]),
+        int("b", [1]),
         bigint("c", [3]),
         bigint("d", [1]),
         varchar("e", ["d"]),
